# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateAccountRequest:
    properties:
      accountName:
        type: string
      accountType:
        type: string
      clientId:
        type: integer
      companyData:
        $ref: '#/definitions/handlers.companyReq'
      createCard:
        type: boolean
      currencyCode:
        type: string
      initialBalance:
        type: number
    required:
    - accountType
    - clientId
    - currencyCode
    type: object
...
  handlers.companyReq:
    properties:
      activityCode:
        type: string
      address:
        type: string
      name:
        type: string
      pib:
        type: string
      registrationNumber:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: REST API gateway for EXBanka microservices.
  title: EXBanka API Gateway
  version: "1.0"
paths:
  /api/accounts:
    get:
      description: Returns all bank accounts. Requires employee authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all accounts
      tags:
      - accounts
  /api/accounts/{accountId}:
    get:
      description: Returns detailed info for a single account. The account must belong
        to the authenticated user.
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account details
      tags:
      - accounts
  /api/accounts/{accountId}/name:
    put:
      consumes:
      - application/json
      description: Changes the name of the account. The account must belong to the
        authenticated user.
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: integer
      - description: New name
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rename account
      tags:
      - accounts
  /api/accounts/create:
    post:
      consumes:
      - application/json
      description: Creates a new bank account for a client. Requires employee authentication.
      parameters:
      - description: Account creation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create bank account
      tags:
      - accounts
  /api/accounts/my:
    get:
      description: Returns all accounts for the currently authenticated client, sorted
        by available balance descending.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my accounts
      tags:
      - accounts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"